pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sqlfmt-web"
version = "0.1.0"
description = "The code behind sqlfmt.com, a web UI for sqlfmt"
authors = ["Ted Conbeer <ted@shandy.io>"]
license = "Apache-2.0"
homepage = "http://sqlfmt.com"
repository = "https://github.com/tconbeer/sqlfmt-web"
packages = [
    { include = "sqlfmt_web", from = "src" }
]

[tool.poetry.dependencies]
python = "3.8.10"
shandy-sqlfmt = "^0.1.0"
pywebio = "^1.4.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^4.0.1"
isort = "^5.9.3"
mypy = "^0.910"
pre-commit = "^2.15.0"
pytest-selenium = "^2.0.1"

[tool.poetry.scripts]
sqlfmt-web = "sqlfmt_web.index:serve_index"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
src_paths = ["src/sqlfmt", "tests"]
